fix undefined

+++ compatibility/rpc/dubbo/go-client/pkg/user.go
@@ -55,7 +55,7 @@ func (g Gender) String() string {
 		return s
 	}
 
-	return strconv.Itoa(int(g))
+	return strconv.Itoa(int(int32(g)))
 }
 
 func (g Gender) EnumValue(s string) hessian.JavaEnum {
